Assignment_4/requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
pywinauto==0.6.8
asyncio==3.4.3 